Assignment 02; Added search input bar which can filter the results in the list of movies.
This still needs some tweaking since the behavoir is a little wonky at times but I will change this in de comming day.

- IntroductionPage.js:
Added text to introduction page.

- MovieItem.js:
Added title to the onClick alert for testing.

- PermDrawer.js:
Removed this file since it was just for testing.

- RouterComponent.js:
Added logic for searching and documentation. Introduced a new state object which will hold the input from the search input (AutoComplete component). Used the Autocomplete component for the search field since this component haves a build in auto complete, this component needs some tweaking. Introduced a new const which will hold the filtered movies based on the input of the Autocomplete component.