Working on displaying the movie Image in the MovieDetails.js.
- MovieDetails.js:
Changed the <Image/> tag to a <Box/> with styling and the appropriate config for displaying the image of the selected movie.

The <Image/> tag was just an Icon, hence it not working to display images.

- RouterComponent.js:
updated assignment doc.

Next step is to create a custom component with error handling so we can display the Icon whenever there is no Image to display due to invalid path or undefined path. The path will be undefined when adding a new movie.