Doing some restructuring by introducing a DataManager that will be responsible for the fetching of the data.
This will make it easier to fetch data within diffrent components which we need for the DetailMovie.js.

The DetailMovie currently relies on it's parent to pass the movie to display where the url param is ignored, resulting in JSON errors when reloading the page. By making the DetailMovie.js fetch it's own data based on the ID this will resolve the issues.

- RouterComponent.js:
Updated the router component such that it utilizes the new DataManager object.

+ DataManager.js:
New object literal that is responsible for the fetching of the Movies data.