Changed the MovieDetails.js such that is also uses the DataManager.js.
Had an issue where only the first render was going well but adding the param id variable to the deps array of the useEffect() fixes the problem.

- MovieItem.js:
Removed now obsolete state property that was given to the useNavigate hook.

- MovieDetails.js:
Removed the unused useLocation() hook since we switched to using the DataManager and now depend on the url param. Changed the useEffect such that is uses the custom DataManager.js.