Again rewriting some things because of "ooh I should do this now to make things easier".
- DataManager.js:
Changed the behavior such that this utilizes the browsers localStorage, so we can have some persistence when saving/changing movies.

- MovieDetails.js:
Due to the changes and implementation of the DataManager.js changes to some logic such as the initial movieState.

The movieState is set to it's initial state whenever there is no id in the url param which indicates this current details screen is loaded because the user clicked the "creation" button. (reuse-ing code)

Added logic to the handleInputChange function since that is needed to be able to change or give input to the input fields. This function will be triggered everytime change is detected to one of the input fields.

Currently workin on the handleSave function... WIP